pyproject.toml
[project]
name = "orchestrator-cc"
version = "0.1.0"
description = "複数のClaude Codeインスタンスを制御するシステム"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "watchdog>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.19.0",
    "ruff>=0.14.0",
    "pytest>=9.0.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.6.0",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]

# === mypy設定 ===
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disable_error_code = ["method-assign"]

# === pytest設定 ===
[tool.pytest.ini_options]
minversion = "9.0"
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "integration: 統合テスト（tmux依存、実環境での実行が必要）",
    "unit: 単体テスト（モック使用、高速実行可能）",
    "slow: 実行に時間がかかるテスト",
]

# === coverage設定 ===
[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/dist/*",
    "*/build/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
fail_under = 80
